src/test/java/it/dcm/bank/filter/LoggingFilterTest.java
package it.dcm.bank.filter;

import jakarta.servlet.FilterChain;
import jakarta.servlet.ServletException;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.test.context.junit.jupiter.SpringExtension;
;
import java.io.IOException;
import java.util.UUID;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.Mockito.*;

@ExtendWith(SpringExtension.class)
class LoggingFilterTest {

    @InjectMocks
    private LoggingFilter loggingFilter;

    @Mock
    private FilterChain filterChain;

    private MockHttpServletRequest request;
    private MockHttpServletResponse response;

    @BeforeEach
    void setUp() {
        request = new MockHttpServletRequest();
        response = new MockHttpServletResponse();
    }

    @Test
    void doFilterInternal_ShouldAddTraceIdToResponse() throws ServletException, IOException {
        loggingFilter.doFilterInternal(request, response, filterChain);

        String correlationId = response.getHeader("X-Trace-Id");
        assertNotNull(correlationId, "Trace ID should not be null");
        assertTrue(UUID.fromString(correlationId).toString().equals(correlationId), "Trace ID should be a valid UUID");

        verify(filterChain).doFilter(request, response);
    }
}
